Add files via upload
These were Randy's comments: Looks pretty good. I don't know that age directly affects HIV status. I would expect it to indirectly modify HIV risk by influencing several other factors that directly modify risk of HIV acquisition. The same could be argued about receiving payment for sex. This is a good start, though. Also, the convention is for a DAG to flow from the top left with the outcome (HIV in our case) in the bottom right. I don't think I emphasized that much in the lecture.

With that being said, I reformatted the DAG and removed the direct connections from age to HIV and payment for sex to HIV.